+++ verb.el
@@ -2086,12 +2086,6 @@ loaded into."
       (verb-util--log num 'W "Body is present but request method is %s"
                       url-request-method))
 
-    ;; Workaround for "localhost" not working on Emacs 25.
-    (when (and (< emacs-major-version 26)
-               (string= (url-host url) "localhost"))
-      (verb-util--log num 'W "Replacing localhost with 127.0.0.1")
-      (setf (url-host url) "127.0.0.1"))
-
     ;; Send the request!
     (condition-case err
         (url-retrieve url